+++ internal/server/handlers/app/settings.go
@@ -1,0 +1,1 @@
+package app
