+++ internal/server/handlers/app/subscriptions.go
@@ -1,0 +1,1 @@
+package app
