--- internal/templates/gen/get_test.go
@@ -1,7 +1,0 @@
-package main
-
-import "testing"
-
-func TestGenerateComponentsTree(t *testing.T) {
-	generateLayoutsTree()
-}
