+++ internal/templates/layouts/main.templ
@@ -4,14 +4,14 @@ import (
   "github.com/byvko-dev/youtube-app/internal/templates/layouts/partials"
 )
 
-templ Main(content templ.Component) {
+templ Main() {
 	<html lang="en" class="h-full">
-		@partials.Head(nil)
+		@partials.Head()
 		<body class="flex flex-col justify-between h-full p-2 mx-auto max-w-7xl" hx-ext="morph">
 			<div id="content" class="flex flex-col flex-grow">
-				{! content }
+				{ children... }
 			</div>
-			@partials.Footer(nil)
+			@partials.Footer()
 		</body>
 	</html>
 }
