+++ internal/templates/layouts/partials/footer.templ
@@ -1,6 +1,6 @@
 package partials
 
-templ Footer(_ templ.Component) {
+templ Footer() {
 	<footer id="footer" class="flex flex-col">
 		<div class="flex flex-row justify-between">
 			<div class="flex flex-col">
