+++ internal/templates/layouts/partials/head.templ
@@ -1,6 +1,6 @@
 package partials
 
-templ Head(content templ.Component) {
+templ Head() {
 	<head>
 		<meta charset="UTF-8"/>
 		<meta name="color-scheme" content="dark"/>
@@ -12,7 +12,7 @@ templ Head(content templ.Component) {
 		<link href="/static/css/daisyui.css" rel="stylesheet" type="text/css"/>
 		<script src="/static/js/tailwind.js"></script>
 		<link rel="manifest" href="/static/app.webmanifest"/>
-		{! content }
+		{ children... }
 	</head>
 }
