+++ internal/templates/pages/error.templ
@@ -1,0 +1,40 @@
+package pages
+
+import (
+  "github.com/byvko-dev/youtube-app/internal/templates/layouts"
+  "github.com/byvko-dev/youtube-app/internal/templates/components/shared"
+)
+
+templ Error(message string) {
+	@layouts.Main() {
+    <div id="error" class="flex flex-col items-center self-center justify-center gap-4 m-auto">
+        <a href="/app">
+            @shared.Logo("128")
+        </a>
+
+        <div class="flex flex-col alert alert-error">
+            <div class="flex flex-row justify-between gap-2">
+                <svg xmlns="http://www.w3.org/2000/svg" class="w-6 h-6 stroke-current shrink-0" fill="none"
+                    viewBox="0 0 24 24">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                        d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z" />
+                </svg>
+                <span class="font-bold">Something went wrong</span>
+            </div>
+            <div class="flex items-start">
+                <span>{ message }</span>
+            </div>
+        </div>
+
+        <div>
+            <a href="/">
+                <button class="btn btn-primary">Home</button>
+            </a>
+            <a href="/">
+                <button class="btn btn-dark">Contact Us</button>
+            </a>
+        </div>
+    </div>
+	}
+}
+
