+++ internal/templates/pages/landing.templ
@@ -1,13 +1,15 @@
 package pages
 
 import (
+  "github.com/byvko-dev/youtube-app/internal/templates/layouts"
   "github.com/byvko-dev/youtube-app/internal/templates/components/shared"
 )
 
 templ Landing() {
+	@layouts.Main() {
   <div class="flex flex-col justify-center h-full gap-4 m-auto" id="landing">
     <div class="flex justify-center">
-      @shared.Logo("256")
+      @shared.Logo("128")
     </div>
     <div class="max-w-4xl text-center">
         <span class="text-xl">
@@ -30,5 +32,6 @@ templ Landing() {
         </a>
     </div>
   </div>
+	}
 }
